make go vet + go lint pass (except big renames)
